Migrate correctlyRounded tests
