+++ src/unittests/f32_interval.spec.ts
@@ -14,7 +14,6 @@ import { objectEquals } from '../common/util/util.js';
 import { kValue } from '../webgpu/util/constants.js';
 import {
   absoluteErrorInterval,
-  correctlyRoundedInterval,
   toF32Interval,
   toF32Matrix,
   toF32Vector,
@@ -24,7 +23,7 @@ import {
   spanF32Intervals,
 } from '../webgpu/util/f32_interval.js';
 import { FPInterval, IntervalBounds } from '../webgpu/util/floating_point.js';
-import { hexToF32, hexToF64, map2DArray, oneULPF32 } from '../webgpu/util/math.js';
+import { hexToF64, map2DArray, oneULPF32 } from '../webgpu/util/math.js';
 
 import { UnitTest } from './unit_test.js';
 
@@ -1636,62 +1635,6 @@ g.test('toF32Matrix')
     t.expect(
       objectEquals(got, expected),
       `toF32Matrix([${input}]) returned [${got}]. Expected [${expected}]`
-    );
-  });
-
-interface CorrectlyRoundedCase {
-  value: number;
-  expected: number | IntervalBounds;
-}
-
-g.test('correctlyRoundedInterval')
-  .paramsSubcasesOnly<CorrectlyRoundedCase>(
-    // prettier-ignore
-    [
-      // Edge Cases
-      { value: kValue.f32.infinity.positive, expected: kAnyBounds },
-      { value: kValue.f32.infinity.negative, expected: kAnyBounds },
-      { value: kValue.f32.positive.max, expected: kValue.f32.positive.max },
-      { value: kValue.f32.negative.min, expected: kValue.f32.negative.min },
-      { value: kValue.f32.positive.min, expected: kValue.f32.positive.min },
-      { value: kValue.f32.negative.max, expected: kValue.f32.negative.max },
-
-      // 32-bit subnormals
-      { value: kValue.f32.subnormal.positive.min, expected: [0, kValue.f32.subnormal.positive.min] },
-      { value: kValue.f32.subnormal.positive.max, expected: [0, kValue.f32.subnormal.positive.max] },
-      { value: kValue.f32.subnormal.negative.min, expected: [kValue.f32.subnormal.negative.min, 0] },
-      { value: kValue.f32.subnormal.negative.max, expected: [kValue.f32.subnormal.negative.max, 0] },
-
-      // 64-bit subnormals
-      { value: hexToF64(0x0000_0000_0000_0001n), expected: [0, kValue.f32.subnormal.positive.min] },
-      { value: hexToF64(0x0000_0000_0000_0002n), expected: [0, kValue.f32.subnormal.positive.min] },
-      { value: hexToF64(0x800f_ffff_ffff_ffffn), expected: [kValue.f32.subnormal.negative.max, 0] },
-      { value: hexToF64(0x800f_ffff_ffff_fffen), expected: [kValue.f32.subnormal.negative.max, 0] },
-
-      // 32-bit normals
-      { value: 0, expected: [0, 0] },
-      { value: hexToF32(0x03800000), expected: hexToF32(0x03800000) },
-      { value: hexToF32(0x03800001), expected: hexToF32(0x03800001) },
-      { value: hexToF32(0x83800000), expected: hexToF32(0x83800000) },
-      { value: hexToF32(0x83800001), expected: hexToF32(0x83800001) },
-
-      // 64-bit normals
-      { value: hexToF64(0x3ff0_0000_0000_0001n), expected: [hexToF32(0x3f800000), hexToF32(0x3f800001)] },
-      { value: hexToF64(0x3ff0_0000_0000_0002n), expected: [hexToF32(0x3f800000), hexToF32(0x3f800001)] },
-      { value: hexToF64(0x3ff0_0010_0000_0010n), expected: [hexToF32(0x3f800080), hexToF32(0x3f800081)] },
-      { value: hexToF64(0x3ff0_0020_0000_0020n), expected: [hexToF32(0x3f800100), hexToF32(0x3f800101)] },
-      { value: hexToF64(0xbff0_0000_0000_0001n), expected: [hexToF32(0xbf800001), hexToF32(0xbf800000)] },
-      { value: hexToF64(0xbff0_0000_0000_0002n), expected: [hexToF32(0xbf800001), hexToF32(0xbf800000)] },
-      { value: hexToF64(0xbff0_0010_0000_0010n), expected: [hexToF32(0xbf800081), hexToF32(0xbf800080)] },
-      { value: hexToF64(0xbff0_0020_0000_0020n), expected: [hexToF32(0xbf800101), hexToF32(0xbf800100)] },
-    ]
-  )
-  .fn(t => {
-    const expected = toF32Interval(t.params.expected);
-    const got = correctlyRoundedInterval(t.params.value);
-    t.expect(
-      objectEquals(expected, got),
-      `correctlyRoundedInterval(${t.params.value}) returned ${got}. Expected ${expected}`
     );
   });
